backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dropbox==11.36.2
ecdsa==0.18.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ply==3.11
pyasn1==0.5.0
pydantic==2.5.0
pydantic_core==2.14.1
pyee==8.2.2
pypng==0.20220715.0
pyppeteer==1.0.2
python-jose==3.3.0
python-multipart==0.0.6
qrcode==7.4.2
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
stone==3.3.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.18
uvicorn==0.24.0.post1
websockets==10.4
zipp==3.17.0
